etl/pyproject.toml
[project]
name = "etl"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
  "boto3~=1.34.142",
  "geopandas==1.0.1",
  "openpyxl>=3.1.5",
  "osm2geojson==0.2.5",
  "pandas==2.2.2",
  "python-dotenv==1.0.1",
  "requests==2.32.3",
]

[tool.setuptools]
packages = []
